Toy: Add initial files
   * Makefile
   * main
   * array tests
   * linked lists

This supports the running of the inital ask for what type of tests
we will want to test.  It'll loop around and do things.  It goes
into the classes and has been manually tested to make sure that
things are running between tne classes and that it can exit
successfully.

This patch includes the necessary headers, and it also adds the node
struct for implementing things in following patches.  It'll allow for
some basic coding toys with it.  Makes for a simple toy for data
structures.

Signed-off-by: Alice Michael <alice.michael@proton.me>